fix(app): 🐛 ensure commit log appears without user interaction
