+++ src/__tests__/appCommitLog.test.tsx
@@ -1,0 +1,38 @@
+/** @jest-environment jsdom */
+import React from 'react';
+import { render, waitFor } from '@testing-library/react';
+import { App } from '../client/App';
+
+const commits = [
+  { commit: { message: 'new', committer: { timestamp: 2 } } },
+  { commit: { message: 'old', committer: { timestamp: 1 } } },
+];
+
+describe('App commit log', () => {
+  const originalFetch = global.fetch;
+
+  beforeEach(() => {
+    jest.resetModules();
+    document.body.innerHTML = '<div id="root"></div>';
+    global.fetch = jest.fn((input: RequestInfo | URL) => {
+      if (typeof input === 'string' && input.startsWith('/api/commits')) {
+        return Promise.resolve({ json: () => Promise.resolve(commits) });
+      }
+      if (typeof input === 'string' && input.startsWith('/api/lines')) {
+        return Promise.resolve({ json: () => Promise.resolve([]) });
+      }
+      return Promise.reject(new Error(`Unexpected url: ${input}`));
+    }) as jest.Mock;
+  });
+
+  afterEach(() => {
+    global.fetch = originalFetch;
+  });
+
+  it('renders commit log after loading', async () => {
+    const { container } = render(<App />);
+    await waitFor(() =>
+      expect(container.querySelectorAll('#commit-log li')).toHaveLength(2),
+    );
+  });
+});
